feat(autocrud): add tests and move style props
Fixes #1640
