--- packages/ts/react-crud/src/shared/types.ts
@@ -1,7 +1,0 @@
-import type { CSSProperties } from 'react';
-
-export type ComponentStyleProps = Readonly<{
-  id?: string;
-  style?: CSSProperties;
-  className?: string;
-}>;
